extract comments into separate components
