+++ src/components/RecipeFullView.jsx
@@ -1,23 +1,12 @@
 import React, { useEffect, useState } from "react";
-import { useParams } from "react-router-dom";
-import {
-  styled,
-  Typography,
-  Grid,
-  Paper,
-  Button,
-  Alert,
-  TextField,
-  Box,
-  Avatar
-} from "@mui/material";
+import { useParams, useNavigate } from "react-router-dom";
+import { styled, Typography, Grid, Paper, Button, Alert } from "@mui/material";
 import Recipes from "../modules/Recipes";
 import IngredientsList from "./IngredientsList";
 import DeleteIcon from "@mui/icons-material/Delete";
-import { useSelector } from "react-redux";
-import { useNavigate } from "react-router-dom";
 import EditIcon from "@mui/icons-material/Edit";
-import Comments from "../modules/Comments";
+import { useSelector } from "react-redux";
+import CommentsSection from "./CommentsSection";
 
 const Img = styled("img")({
   margin: "auto",
@@ -33,7 +22,6 @@ const RecipeFullView = () => {
   const [recipe, setRecipe] = useState({});
   const [showEditDelete, setShowEditDelete] = useState(false);
   const [message, setMessage] = useState();
-  const [comment, setComment] = useState();
   const [commentsList, setCommentsList] = useState([]);
 
   const fetchRecipe = async () => {
@@ -43,45 +31,6 @@ const RecipeFullView = () => {
       setRecipe(data.recipe);
       setCommentsList(data.recipe.comments);
     }
-  };
-
-  const commentsFeed = commentsList.map((comment, index) => {
-    comment.index = index + 1;
-    return (
-      <Grid item xs={12} key={comment.index}>
-        <Typography
-          data-cy={`comment-user-${comment.index}`}
-          variant="h6"
-          gutterBottom
-          component="div"
-        >
-          {comment.user}
-        </Typography>
-        <Typography
-          data-cy={`comment-body-${comment.index}`}
-          variant="body1"
-          gutterBottom
-          component="div"
-        >
-          {comment.body}
-        </Typography>
-      </Grid>
-    );
-  });
-
-  const createComment = async (event) => {
-    const response = await Comments.create(id, comment);
-    if (response.comment) {
-      let newComment = { user: currentUser.name, body: response.comment?.body }
-      setCommentsList([newComment, ...commentsList]);
-    }
-  };
-
-  const handleChange = (event) => {
-    setComment({
-      ...comment,
-      [event.target.name]: event.target.value
-    });
   };
 
   const deleteRecipe = async () => {
@@ -176,35 +125,10 @@ const RecipeFullView = () => {
           </Grid>
         </Grid>
       </Paper>
-      <Box
-        component="form"
-        sx={{ p: 2, margin: "auto", maxWidth: 800, flexGrow: 1, boxShadow: 3 }}
-        noValidate
-        autoComplete="off"
-      >
-        {currentUser && (
-          <React.Fragment>
-            <TextField
-              data-cy="comment-field"
-              name="comment"
-              size="normal"
-              variant="outlined"
-              fullWidth
-              multiline
-              placeholder="Leave your comment here ..."
-              onChange={handleChange}
-            />
-            <Button onClick={createComment} data-cy="post-comment-btn">
-              Post comment
-            </Button>
-          </React.Fragment>
-        )}
-        <Paper style={{ padding: "40px 20px" }}>
-          <Grid container data-cy="comment-feed" spacing={2}>
-            {commentsFeed}
-          </Grid>
-        </Paper>
-      </Box>
+      <CommentsSection
+        onCommentAdded={fetchRecipe}
+        commentsList={commentsList}
+      />
     </React.Fragment>
   );
 };
